+++ mixout/mixlinear.py
@@ -7,13 +7,12 @@
 ##++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 import math
 import torch
-from torch import nn
 import torch.nn.init as init
 import torch.nn.functional as F
 from torch.nn import Parameter
-from mixout import mixout
 from typing import Optional
 from collections import OrderedDict
+from functional import mixout
